ssl_session_cache_api behaviour (ssl v11.0.3)

TLS session cache API

Defines the API for the TLS session cache (pre TLS-1.3) so that the data storage scheme can be replaced by defining a new callback module implementing this API.

Types: Types used in TLS/DTLS

reason()
(since OTP R14B)

-type reason() :: any().

Types

cert_key/0
(since OTP R14B)

-type cert_key() :: {reference(), integer(), issuer()}.

certdb_ref/0
(since OTP R14B)

-type certdb_ref() :: reference().

db_handle/0
(since OTP R14B)

-type db_handle() :: any().

der_cert/0
(since OTP R14B)

-type der_cert() :: binary().

from/0
(since OTP R14B)

-type from() :: any().

gen_fsm_state_return/0
(since OTP R14B)

-type gen_fsm_state_return() ::
    {next_state, state_name(), any()} |
    {next_state, state_name(), any(), timeout()} |
    {stop, any(), any()}.

issuer/0
(since OTP R14B)

-type issuer() :: tuple().

msg/0
(since OTP R14B)

-type msg() :: any().

partial_key()
(opaque) (since OTP R14B)

-opaque partial_key()

The opaque part of the key. Does not need to be handled by the callback.

reply/0
(since OTP R14B)

-type reply() :: any().

secret_printout/0
(since OTP R14B)

-type secret_printout() :: list().

serialnumber/0
(since OTP R14B)

-type serialnumber() :: integer().

session()
(opaque) (since OTP R14B)

-opaque session()

The session data that is stored for each session.

session_cache_key()
(since OTP R14B)

-type session_cache_key() :: {partial_key(), ssl:session_id()}.

A key to an entry in the session cache.

session_cache_ref()
(since OTP R14B)

-type session_cache_ref() :: any().

ssl_options/0
(since OTP R14B)

-type ssl_options() :: map().

state_name/0
(since OTP R14B)

-type state_name() :: hello | abbreviated | certify | cipher | connection.

Callbacks

delete/2
(since OTP R14B)

-callback delete(session_cache_ref(), session_cache_key()) -> any().

Deletes a cache entry. Is only called from the cache handling process.

foldl/3
(optional) (since OTP R14B)

-callback foldl(fun(), term(), session_cache_ref()) -> term().

Calls Fun(Elem, AccIn) on successive elements of the cache, starting with AccIn == Acc0. Fun/2 must return a new accumulator, which is passed to the next call. The function returns the final value of the accumulator. Acc0 is returned if the cache is empty.

Note

Since OTP-23.3 this functions is only used on the client side and does not need to implemented for a server cache.

init/1
(since OTP 18.0)

-callback init(list()) -> session_cache_ref().

Includes property {role, client | server}. Currently this is the only predefined property, there can also be user-defined properties. See also application environment variable session_cb_init_args.

Performs possible initializations of the cache and returns a reference to it that is used as parameter to the other API functions. Is called by the cache handling processes init function, hence putting the same requirements on it as a normal process init function. This function is called twice when starting the SSL application, once with the role client and once with the role server, as the SSL application must be prepared to take on both roles.

lookup/2
(since OTP R14B)

-callback lookup(session_cache_ref(), session_cache_key()) -> #session{} | undefined.

Looks up a cache entry. Is to be callable from any process.

select_session/2
(optional) (since OTP R14B)

-callback select_session(session_cache_ref(), {ssl:host(), inet:port_number()} | inet:port_number()) ->
                  [#session{}].

Selects sessions that can be reused, that is sessions that include PartialKey in its key. Is to be callable from any process.

Note

Since OTP-23.3 This functions is only used on the client side and does not need to implemented for a server cache.

size/1
(since OTP 19.3)

-callback size(session_cache_ref()) -> integer().

Returns the number of sessions in the cache. If size exceeds the maximum number of sessions, the current cache entries will be invalidated regardless of their remaining lifetime. Is to be callable from any process.

terminate/1
(since OTP R14B)

-callback terminate(session_cache_ref()) -> any().

Takes care of possible cleanup that is needed when the cache handling process terminates.

update/3
(since OTP R14B)

-callback update(session_cache_ref(), session_cache_key(), #session{}) -> any().

Caches a new session or updates an already cached one. Is only called from the cache handling process.